+++ src/NEventStore.Benchmark/Support/SomeDomainEvent.cs
@@ -1,5 +1,6 @@
-﻿namespace NEventStore.Benchmark.Support
+namespace NEventStore.Benchmark.Support
 {
+    [Serializable]
     internal class SomeDomainEvent
     {
         public string? Value { get; set; }